requirements.txt
bidict==0.23.1
blinker==1.8.2
boto3==1.35.76
botocore==1.35.76
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
config==0.5.1
cryptography==43.0.3
deepl==1.20.0
Flask==3.0.3
Flask-SocketIO==5.4.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
s3transfer==0.10.4
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==1.26.16
Werkzeug==3.0.6
wsproto==1.2.0
gunicorn
flask_cors
flask
pymysql
python-dotenv
google-api-python-client
oauth2client
